Default super().__torch_dispatch__ to one hop dispatch if composite available
Today, if you have a CompositeImplicitAutograd kernel but it didn't get
applied (because, e.g., inference mode was enabled), once you get to the
`__torch_dispatch__` handler you have no way of saying "well actually, I
wanted the composite".  This patch makes super().__torch_dispatch__, by
default, attempt a decomposition (if it exists), before permanently
disabling the Python keys and running the conventional kernel.

The way it works is copying the same strategy as
https://github.com/pytorch/pytorch/pull/74728 for "skip one hop".
Instead of adding Python to the exclude set (which will prevent us from
dispatching back to this key), we set a boolean TLS that says "don't
process the next Python key.)  When we see this skip bool is set,
we set it to false, and redispatch past the Python key.  This will take
you to the newly added PythonComposite key, which will either (1) run
the CompositeImplicitAutograd decomposition, with the redispatches
calling into Python, because the skip bool is no longer set, or (2)
exclude Python/PythonComposite permanently and drop into the true
backend implementation.

Not in this PR yet, but PythonComposite offers a last chance for us to
check for wrapper subclasses before going into backend code.  Wrapper
subclasses cannot be validly operated on by, e.g., CPU/CUDA kernels, so
erroring here could help prevent segfaults.  There is one part missing
from this PR, which is if you have a tensor subclass that isn't supported
in the backend, how can you tell if it is safe to attempt the composite
or not?  Right now you can just ask if there is a
CompositeImplicitAutograd kernel registered for this operator, but this
is a bit low level.

We discussed a number of alternate strategies, which also seem viable.
This one just seemed the simplest to implement.

- We could make it possible to directly call a kernel at a specific
  dispatch key from Python.  In this case, you could test if
  CompositeImplicitAutograd kernel exists, and if so, call it directly.
  I didn't like this because (1) when you call the kernel directly, you
  are obligated to say what the dispatchKeySet it was called at is, but
  this is undefined if you're directly calling a kernel, and (2) it
  seems like a breakage of encapsulation; normally you have to go
  through the dispatcher to work out what kernel to call next, and this
  lets you bypass it entirely.

- We could directly make the C++ composite functions available from
  Python, and then call them directly when we realize they're available.
  This is annoying to implement since operator registration is not fixed
  (it can happen on dynamic library load) so you need a full listener
  infrastructure to do this properly.

Signed-off-by: Edward Z. Yang <ezyangfb.com>

ghstack-source-id: b226d8884e45f95efdba0fa24243687c85067659
Pull Request resolved: https://github.com/pytorch/pytorch/pull/82976