Respect TorchDispatchMode for shallow_copy_and_detach
I noticed I was missing tensor creations with modes when I tried
to delete proxy tensor.  This was the cause.

Hypothetically, all PyInterpreter calls could get this treatment.
But I think it only matters for detach; the rest do not return
Tensors and most modes will not be interested in them.

Signed-off-by: Edward Z. Yang <ezyangfb.com>

ghstack-source-id: 49cfc613656358fe1dc1436a2a6f7b8541a8edf8
Pull Request resolved: https://github.com/pytorch/pytorch/pull/83372