[torchgen] Add logic in annotation parser to accept alias set
Extending the current regex in `model.py` to support annotation alias set. See issue #83214.

Ideally we should have a full fledged lexer similar to `schema_type_parser.cpp`, since regex can be more and more difficult to read if we add more support to it.

Adding this to unblock this issue for now.

ghstack-source-id: e7d0c025a1b131494a323dd4d5bedb9f9a63e7d4
Pull Request resolved: https://github.com/pytorch/pytorch/pull/83501