[vulkan] Throw c10::Error instead of using TORCH_CHECK when checking VkResult return codes
Pull Request resolved: https://github.com/pytorch/pytorch/pull/83627

Currently the `VkResult` of Vulkan API is checked with `TORCH_CHECK(...)`.

This diff changes it to throw a `c10::Error` instead. The reason for this is that `TORCH_CHECK()` will not preserve error messages `#ifdef STRIP_ERROR_MESSAGES`. However, it is more useful for the error reason to be always present for the purpose of diagnosing driver support issues and detecting if a model load failure is related to Vulkan.

This diff also removes incorrect error throwing when initializing the global context/runtime.

ghstack-source-id: 164998243

Differential Revision: [D38800348](https://our.internmc.facebook.com/intern/diff/D38800348/)