Refactor is_X_like, better invariant checking for SymInt overload
Add is_symint_like, by way of is_base_ty_like which generalizes
the pattern for is_tensor_like and is_generator_like.  Now that
we can query if a signature contains a SymInt, we can enforce that
you must name the overload with SymInt if the signature contains
SymInt.

Signed-off-by: Edward Z. Yang <ezyangfb.com>

ghstack-source-id: 481d8e99ab2ceb9bef6f897a72775185393f5513
Pull Request resolved: https://github.com/pytorch/pytorch/pull/83668